requirements.txt
torch
torchvision
pygame
numpy
pillow